% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHandDistribution.R
\name{createHandDistribution}
\alias{createHandDistribution}
\title{createHandDistribution}
\usage{
createHandDistribution(NoSimulations)
}
\arguments{
\item{NoSimulations}{{ number of simulations to perform}}
}
\value{
simhandTibble { a tibble with the simulations}
}
\description{
createHandDistribution
}
